refactor(frontend): サービス層ファイル名をケバブケースへ統一
